Section-27: converted api's to helper functions and externalized them to make the server file simpler.
